refactor : change directories based on FSD
