+++ src/app/main.jsx
@@ -3,7 +3,7 @@ import { createRoot } from 'react-dom/client';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import HomePage from '../pages/HomePage.jsx';
 import AppFrame from '../layouts/AppFrame.jsx';
-import RestaurantDetailPage from '../components/RestaurantDetailPage.jsx';
+import RestaurantDetailPage from '../pages/RestaurantDetailPage.jsx';
 import RestaurantDetailFrame from '../layouts/RestaurantDetailFrame.jsx';
 import ErrorPage from '../pages/ErrorPage.jsx';
