+++ src/pages/HomePage.jsx
@@ -1,9 +1,9 @@
 import styled from '@emotion/styled';
 import rice from '../assets/rice.svg';
 import kiryong from '../assets/kiryong-notice.svg';
-import MenuCardSection from '../components/MenuCardSection';
+import MenuCardSection from './MenuCardSection';
 import { getTodayDate } from '../utils/getTodayDate';
-import InnerTitlesWrapper from '../components/InnerTitlesWrapper';
+import InnerTitlesWrapper from '../widgets/InnerTitlesWrapper';
 
 export default function Home() {
   return (
