+++ src/shared/utils/fetchData.js
